moving aggregationISM outside testnet3/
