+++ typescript/infra/config/environments/testnet3/core.ts
@@ -5,13 +5,14 @@ import {
   objMap,
 } from '@hyperlane-xyz/sdk';
 
+import { aggregationIsm } from '../../aggregationIsm';
 import { Contexts } from '../../contexts';
 
-import { aggregationIsm } from './aggregationIsm';
 import { owners } from './owners';
 
 export const core: ChainMap<CoreConfig> = objMap(owners, (local, owner) => {
   const defaultIsm: AggregationIsmConfig = aggregationIsm(
+    'testnet3',
     local,
     Contexts.Hyperlane,
   );
